backend/db/feature4/requirements.txt
pdfplumber==0.11.4
pymupdf==1.24.11
sentence-transformers==3.3.1
pinecone==5.0.1
supabase==2.7.4
openai>=1.30.0
fpdf2>=2.8.1
python-dotenv==1.0.1
tqdm==4.66.5
